+++ resources/META-INF/extensions/inspections/latex/code-maturity.xml
@@ -24,9 +24,5 @@
                          groupPath="LaTeX" groupName="Code maturity" displayName="Discouraged use of \makeatletter in tex sources"
                          enabledByDefault="true"
                          level="WARNING" />
-        <localInspection language="Latex" implementationClass="nl.hannahsten.texifyidea.inspections.latex.codematurity.LatexMissingGlossaryReferenceInspection"
-                         groupPath="LaTeX" groupName="Code maturity" displayName="Missing glossary reference"
-                         enabledByDefault="true"
-                         level="WARNING" />
     </extensions>
 </idea-plugin>
