+++ test/nl/hannahsten/texifyidea/inspections/latex/typesetting/LatexMissingGlossaryReferenceInspectionTest.kt
@@ -1,0 +1,23 @@
+package nl.hannahsten.texifyidea.inspections.latex.typesetting
+
+import nl.hannahsten.texifyidea.file.LatexFileType
+import nl.hannahsten.texifyidea.inspections.TexifyInspectionTestBase
+
+class LatexMissingGlossaryReferenceInspectionTest : TexifyInspectionTestBase(LatexMissingGlossaryReferenceInspection()) {
+
+    fun testMissingReference() {
+        myFixture.configureByText(LatexFileType, """\newglossaryentry{sample}{name={sample},description={an example}} \gls{sample} \Glslink{sample} <warning descr="Missing glossary reference">sample</warning>""")
+        myFixture.checkHighlighting()
+    }
+
+    fun testAddGls() {
+        testQuickFix(
+            """
+                \newglossaryentry{sample}{name={sample},description={an example}} \gls{sample} sample text
+            """.trimIndent(),
+            """
+                \newglossaryentry{sample}{name={sample},description={an example}} \gls{sample} \gls{sample} text
+            """.trimIndent()
+        )
+    }
+}
